scripts/requirements.txt
requests>=2.31.0
feedparser>=6.0.10
python-dateutil>=2.8.2
praw>=7.7.1
